U3: add comments and extract tree implementation
